move subcomponent hooks back into their parent files
