+++ client/src/components/Today/Task.tsx
@@ -1,12 +1,46 @@
-import useTask from "@/components/Today/hooks/use-task";
+import { useDetailedActivityModal } from "@/components/Today/hooks/use-detailed-activity-modal";
+import { filterTagsById } from "@/lib/filter-tags";
+import usePutTaskCompletion from "@/lib/hooks/use-put-task-completion";
 import { Checkbox } from "@/lib/theme/components/Checkbox";
 import { activityEnd, activityStart } from "@lib/activity";
 import type { ActivityWithIds } from "@type/server/activity.types";
 import type { TagWithIds } from "@type/server/tag.types";
-import type { ById } from "@type/server/utility.types";
+import type { ById, ID } from "@type/server/utility.types";
+import { useCallback, useRef } from "react";
 import TagCard from "../TagCard/TagCard";
 import T from "./style/Tasks.style";
 import S from "./style/Today.style";
+
+function useTask({
+	activity,
+	tagsById
+}: {
+	activity: ActivityWithIds;
+	tagsById?: Record<ID, TagWithIds>;
+}) {
+	const tags = filterTagsById(activity.tag_ids, tagsById); // TODO: pass this as a prop and, in Tasks, get it from a hook
+	const checkboxRef = useRef<HTMLLabelElement>(null);
+
+	const putCompletion = usePutTaskCompletion(activity);
+
+	const { openDetailedActivityModal } = useDetailedActivityModal({ activity });
+	const maybeOpenTaskModal = useCallback(
+		(e: React.MouseEvent) => {
+			if (checkboxRef.current?.contains(e.target as Node)) return;
+
+			openDetailedActivityModal();
+			e.stopPropagation();
+		},
+		[checkboxRef, activity]
+	);
+
+	return {
+		checkboxRef,
+		maybeOpenTaskModal,
+		putCompletion,
+		tags
+	} as const;
+}
 
 type TaskProps = {
 	activity: ActivityWithIds;
